requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.3
asyncmy==0.2.8
attrs==23.1.0
charset-normalizer==3.2.0
discord.py==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
multidict==6.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
python-osrsapi==0.0.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
yarl==1.9.2
